Simple implementation of MakeCredential response
* Software based ECDSA implementation
* Single hardcoded private key

+++ include/util.h
@@ -5,4 +5,4 @@
 #define MIN(a, b) (((a) < (b)) ? (a) : (b))
 #define MAX(a, b) (((a) > (b)) ? (a) : (b))
 
-void serialDumpBuffer(uint8_t *buffer, uint16_t len);
+void serialDumpBuffer(uint8_t *buffer, size_t len);
