+++ src/util/util.cpp
@@ -5,7 +5,7 @@
 /**
  * @brief Dump contents of the given buffer to the serial output
  */
-void serialDumpBuffer(uint8_t *buffer, uint16_t len)
+void serialDumpBuffer(uint8_t *buffer, size_t len)
 {
     const uint16_t pageSize = 40;
     uint16_t pages = (len / pageSize) + ((len % pageSize) ? 1 : 0);
